requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-yasg==1.21.9
et_xmlfile==2.0.0
folium==0.19.5
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg==3.2.5
psycopg-binary==3.2.5
pydantic==2.10.6
pydantic_core==2.27.2
pyogrio==0.10.0
pyproj==3.7.1
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
shapely==2.0.7
six==1.17.0
sqlparse==0.5.3
structlog==25.1.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
xyzservices==2025.1.0
